requirements.txt
requests
beautifulsoup4
readability-lxml
lxml
notion-client
python-dotenv
lmstudio